requirements.txt
sentence-transformers
faiss-cpu
numpy
transformers
torch
nltk
PyMuPDF
tqdm

#python data/import_pdfs.py
#python build_embeddings.py